requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
WTForms==3.1.2
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0
Pillow>=10.2.0
pytesseract==0.3.10
matplotlib>=3.8.0
seaborn>=0.13.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
python-dotenv==1.0.1
python-dateutil==2.9.0
gunicorn==22.0.0
openai>=1.12.0
google-generativeai>=0.8.0
openpyxl==3.1.5